requirements.txt
fastapi>=0.129.0
uvicorn>=0.30.6
python-dotenv>=0.20.0
redis>=5.0.0
pymongo>=4.16.0
motor>=3.3.0
google-generativeai>=0.8.6
numpy>=1.25.0
tiktoken>=0.12.0
transformers>=5.2.0
requests>=2.31.0
httpx>=0.27.0
python-multipart>=0.0.12
pydantic>=2.0.0
json-repair>=0.28.0
python-dateutil>=2.8.2
sentence-transformers>=3.0.0
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
